+++ src/main/kotlin/racingCar/domain/RacingCar.kt
@@ -1,4 +1,7 @@
-package racingCar
+package racingCar.domain
+
+import racingCar.view.InputView
+import racingCar.view.OutputView
 
 const val RANDOM_LIMIT_NUM = 10
 
@@ -31,13 +34,4 @@ class RacingCar(private val carNames: List<String>) {
         }
         return carsInRacing
     }
-}
-
-fun main() {
-    val carNames = InputView.getCarName()
-    val trialCount = InputView.getTrialCount()
-    val racing = RacingCar(carNames)
-    val resultRacingCar = racing.run { racing.carMove(trialCount) }
-    val returnWinner = Winner.getRacingWinner(resultRacingCar)
-    OutputView.getCarRacingWinners(returnWinner)
 }
