+++ src/main/kotlin/racingCar/domain/Winner.kt
@@ -1,4 +1,4 @@
-package racingCar
+package racingCar.domain
 
 object Winner {
