+++ src/main/kotlin/racingCar/view/InputView.kt
@@ -1,4 +1,4 @@
-package racingCar
+package racingCar.view
 
 import java.lang.IllegalArgumentException
